Add Python makefile
This replaces the traditional makefile with a short Python script.
This also uses pkg-config to find the location of the libxmp header,
so it doesn't depend on having a clone of the libxmp repository in
another path.

Finally, this fixes up the sed command so that it works on BSD sed.