Fix play_buffer implementation and add callback example
Add callback example to exercise play_buffer. This method must return
both the data buffer and an module end signaling. Sometimes a callback
passes the buffer to be filled so it's good to have it passed as a
parameter, so we return a boolean to tell if we're finished. Also
create a static method buffer() to create the buffer in a format
appropriate for play_buffer().

TODO: check if this call format makes sense. I'm not completely happy
      with it, perhaps there's a better way.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>